% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg.R
\name{bg}
\alias{bg}
\title{Generate bias raster (Kernel Smoothed Intensity of Point Pattern, density.ppp) and/or corresponding background points}
\usage{
bg(
  p,
  r,
  n = 0.1,
  sigma = c(0.1, 1, 2, 3),
  output = c("br", "bg"),
  crs = NA,
  anisotropic = FALSE
)
}
\arguments{
\item{p}{sf/sfc (POINT/MULTIPOINT): Points}

\item{r}{RasterLayer: Template raster}

\item{n}{integer or numeric: Ratio (< 1) or int to sample.}

\item{sigma}{numeric vector: \code{density.ppp} \emph{sigma} parameter, multiplies \emph{r} pixel size \emph{x} and/or \emph{y} resolution (for isotropic or anisotropic kernel)}

\item{output}{character vector: \emph{br} (bias raster) and/or \emph{bg} (corresponding background points generated by \code{tgbg::sample})}

\item{crs}{integer: Force crs.}
}
\value{
named list: \emph{br}: RasterLayer (bias raster); \emph{bg}: sf (POINT/MULTIPOINT) (background points) with nested List of choosen \emph{sigma} levels
}
\description{
Generate bias raster (Kernel Smoothed Intensity of Point Pattern, density.ppp) and/or corresponding background points
}
