% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample}
\alias{sample}
\title{Sample from raster or points}
\usage{
sample(rp, n = 0.1, prob = FALSE, crs = NA)
}
\arguments{
\item{rp}{RasterLayer or sf/sfc (POINT/MULTIPOINT): Raster or points.}

\item{n}{integer or numeric: Ratio (< 1) or int to sample.}

\item{prob}{logical: For raster with probabilities as pixel values only.}

\item{crs}{integer: Force crs.}
}
\value{
sf (POINT/MULTIPOINT)
}
\description{
Sample from raster or points
}
\references{
Inspired by: https://github.com/danlwarren/ENMTools/blob/004a4a1e182127900a5f62bc015770479bcd0415/R/check.bg.R#L138-L144
}
