% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgobVersions.R
\name{tgobVersions}
\alias{tgobVersions}
\title{Subsample versions (TO, TS, ssosTO, ssosTS, ssosTGOB) from TGOB}
\usage{
tgobVersions(
  p,
  r = NA,
  species = "species",
  TS.n = 0.1,
  observers = NA,
  TO.n = 0.1,
  observersRemoveSingle = TRUE,
  badWordsSpecies = NA,
  badWordsObservers = NA,
  crs = NA
)
}
\arguments{
\item{p}{sf/sfc (POINT/MULTIPOINT): Points}

\item{r}{RasterLayer: Template raster. If provided, all stats are calculated per raster pixel (cell/square).}

\item{species}{character: column name with species name}

\item{TS.n}{integer or numeric: Ratio (< 1) or int to select for TOP species}

\item{observers}{character: column name with observers' names (expected: "John Doe" / "Doe J." / "John Doe, Jane Maria Moe" / "Doe J., Moe J. M." / ...)}

\item{TO.n}{integer or numeric: Ratio (< 1) or int to select for TOP observers}

\item{observersRemoveSingle}{logical: remove single word observers}

\item{badWordsSpecies}{character vector: observers containing unwanted strings}

\item{badWordsObservers}{character vector: observers containing unwanted strings}

\item{crs}{integer: Force crs.}
}
\value{
sf (POINT/MULTIPOINT): inserted \emph{p} with new (0/1) columns: TO, TS, ssosTGOB_\emph{species}, ssosTO_\emph{species}, ssosTS_\emph{species}
}
\description{
Subsample versions (TO, TS, ssosTO, ssosTS, ssosTGOB) from TGOB
}
