% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgobVersions.R
\name{tgobVersions}
\alias{tgobVersions}
\title{Subsample versions (TO, TS and ssosX_\emph{species}) from TGOB}
\usage{
tgobVersions(
  p,
  r = NA,
  species = "species",
  TS.n = 0.2,
  observers = NA,
  TO.n = 0.2,
  observersRemoveSingleName = TRUE,
  observersRemoveSingleOccurrence = 1,
  quantileThreshold = 0.01,
  badWordsSpecies = NA,
  badWordsObservers = NA,
  crs = NA,
  prefix = "nc_"
)
}
\arguments{
\item{p}{sf/sfc (POINT/MULTIPOINT): Points}

\item{r}{RasterLayer: Template raster. If provided, all stats are calculated per raster pixel (cell/square).}

\item{species}{character: column name with species name}

\item{TS.n}{integer or numeric: Ratio (< 1) or int to select for TOP species}

\item{observers}{character: column name with observers' names (expected: "John Doe" / "Doe J." / "John Doe, Jane Maria Moe" / "Doe J., Moe J. M." / ...)}

\item{TO.n}{integer or numeric: Ratio (< 1) or int to select for TOP observers}

\item{observersRemoveSingleName}{logical: remove single word (name) observers}

\item{observersRemoveSingleOccurrence}{integer: remove observers with single (or specified number) species occurrence from ssos. Can be too restrictive for very rare species, where observers have small chance to re-observe. On the other hand, can be useful to restrict very active observers that record some species only purposefully.}

\item{quantileThreshold}{numeric: 0-1 (centile 0.01, median 0.5, ...)}

\item{badWordsSpecies}{character vector: observers containing unwanted strings}

\item{badWordsObservers}{character vector: observers containing unwanted strings}

\item{crs}{integer: Force crs.}

\item{prefix}{character: new columns prefix with versions to select automatically}
}
\value{
List: \emph{t}: sf (POINT/MULTIPOINT): inserted \emph{p} with new (0/1) columns: TO, TS, ssosX_\emph{species}; \emph{report}: TO+TS stats uset to treshold selected top O+S
}
\description{
Subsample versions (TO, TS and ssosX_\emph{species}) from TGOB
}
